Move custom rules/messages to custom.go
