+++ internal/pkg/validator/custom.go
@@ -1,0 +1,90 @@
+package validator
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/go-playground/validator/v10"
+	"github.com/umisama/go-regexpcache"
+
+	"github.com/keboola/keboola-as-code/internal/pkg/utils/strhelper"
+)
+
+func (v *wrapper) registerCustomMessages() {
+	v.registerErrorMessage("required_if", "{0} is a required field")
+}
+
+func (v *wrapper) registerCustomRules() {
+	v.RegisterRule(
+		// Register default validation for "required_in_project"
+		// Some values are required in the project scope, but ignored in the template scope.
+		// We validate them by default.
+		Rule{
+			Tag: "required_in_project",
+			FuncCtx: func(ctx context.Context, fl validator.FieldLevel) bool {
+				if v, _ := ctx.Value(DisableRequiredInProjectKey).(bool); v {
+					// Template mode, value is valid.
+					return true
+				}
+				// Project mode, value must be set.
+				return !fl.Field().IsZero()
+			},
+			ErrorMsg: "{0} is a required field",
+		},
+		Rule{
+			Tag: "required_not_empty",
+			Func: func(fl validator.FieldLevel) bool {
+				// Check nil
+				if !fl.Field().IsValid() {
+					return false
+				}
+				// Check empty
+				return fl.Field().Len() > 0
+			},
+			ErrorMsg: "{0} is a required field",
+		},
+		// Alphanumeric string with allowed dash character.
+		Rule{
+			Tag: "alphanumdash",
+			FuncCtx: func(ctx context.Context, fl validator.FieldLevel) bool {
+				return regexpcache.MustCompile(`^[a-zA-Z0-9\-]+$`).MatchString(fl.Field().String())
+			},
+			ErrorMsg: "{0} can only contain alphanumeric characters and dash",
+		},
+		// Template icon
+		Rule{
+			Tag: "templateicon",
+			FuncCtx: func(ctx context.Context, fl validator.FieldLevel) bool {
+				value := fl.Field().String()
+				if strings.HasPrefix(value, "component:") {
+					value = strings.TrimPrefix(value, "component:")
+					return len(value) > 0
+				}
+				if strings.HasPrefix(value, "common:") {
+					value = strings.TrimPrefix(value, "common:")
+					return allowedTemplateIcons[value]
+				}
+				return false
+			},
+			ErrorMsg: "{0} does not contain an allowed icon",
+		},
+		Rule{
+			Tag: "mdmax",
+			FuncCtx: func(ctx context.Context, fl validator.FieldLevel) bool {
+				value := fl.Field().String()
+				paramStr := fl.Param()
+				param, err := strconv.Atoi(paramStr)
+				if err != nil {
+					panic(fmt.Sprintf("failed to convert mdmax param \"%v\" to an int", paramStr))
+				}
+				value = strhelper.StripMarkdown(value)
+				return len(value) <= param
+			},
+			ErrorMsgFunc: func(fe validator.FieldError) string {
+				return fmt.Sprintf("%s exceeded maximum length of %s", fe.Tag(), fe.Param())
+			},
+		},
+	)
+}
